Switched tablayout to favorites view
