+++ app/src/main/java/com/minjee/favoritepokemon/application/KoinModule.kt
@@ -2,7 +2,8 @@ package com.minjee.favoritepokemon.application
 
 import com.minjee.favoritepokemon.data.NetworkApi
 import com.minjee.favoritepokemon.data.PokemonRepository
-import com.minjee.favoritepokemon.ui.pokemon.allpokemon.AllPokemonTabViewModel
+import com.minjee.favoritepokemon.ui.favorites.allpokemon.AllPokemonTabViewModel
+import com.minjee.favoritepokemon.ui.pokemon.PokemonViewModel
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -12,6 +13,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 val viewModelModule = module {
     viewModel { AllPokemonTabViewModel(get()) }
+    viewModel { PokemonViewModel(get()) }
     single { PokemonRepository(get()) }
 }
