+++ app/src/main/java/com/minjee/favoritepokemon/ui/favorites/FavoritesFragment.kt
@@ -5,35 +5,62 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.FragmentActivity
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.google.android.material.tabs.TabLayoutMediator
+import com.minjee.favoritepokemon.R
 import com.minjee.favoritepokemon.databinding.FragmentFavoritesBinding
+import com.minjee.favoritepokemon.ui.favorites.allpokemon.AllPokemonTabFragment
+import com.minjee.favoritepokemon.ui.favorites.mypokemon.MyPokemonTabFragment
 
 class FavoritesFragment : Fragment() {
-
-    private lateinit var favoritesViewModel: FavoritesViewModel
-
-    private var _binding: FragmentFavoritesBinding? = null
-    private val binding get() = _binding!!
+    private var binding: FragmentFavoritesBinding? = null
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        favoritesViewModel =
-            ViewModelProvider(this).get(FavoritesViewModel::class.java)
-        _binding = FragmentFavoritesBinding.inflate(inflater, container, false)
+        val binding = FragmentFavoritesBinding.inflate(inflater, container, false)
+        this.binding = binding
+
+        val viewPager = binding.workViewPager
+        viewPager.adapter = ViewPagerAdapter(activity)
+
+        val tabLayout = binding.fragmentPokemonTabLayout
+        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
+            tab.text = (viewPager.adapter as ViewPagerAdapter?)!!.tabFragmentTitles[position]
+        }.attach()
+
         return binding.root
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.textDashboard.text = "fave pokemon fragment teksti"
+    override fun onDestroyView() {
+        binding = null
+        super.onDestroyView()
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
+    class ViewPagerAdapter(activity: FragmentActivity?) : FragmentStateAdapter(activity!!) {
+
+        val tabFragmentTitles = arrayOf(
+            activity?.getString(R.string.fragment_tab_all_pokemon_title),
+            activity?.getString(R.string.fragment_tab_my_pokemon_title)
+        )
+
+        override fun getItemCount(): Int {
+            return tabFragments.size
+        }
+
+        override fun getItemId(position: Int): Long {
+            return super.getItemId(position)
+        }
+
+        override fun createFragment(position: Int): Fragment {
+            return tabFragments[position]
+        }
+
+        private val tabFragments = arrayOf(
+            AllPokemonTabFragment(),
+            MyPokemonTabFragment()
+        )
     }
 }
