+++ app/src/main/java/com/minjee/favoritepokemon/ui/favorites/allpokemon/AllPokemonTabFragment.kt
@@ -1,6 +1,5 @@
-package com.minjee.favoritepokemon.ui.pokemon.allpokemon
+package com.minjee.favoritepokemon.ui.favorites.allpokemon
 
-import android.R
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -8,13 +7,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.minjee.favoritepokemon.data.PokemonList
 import com.minjee.favoritepokemon.databinding.FragmentTabAllPokemonBinding
-import com.minjee.favoritepokemon.ui.pokemon.details.PokemonDetailsFragment
 import com.minjee.favoritepokemon.ui.recyclerview.PokemonRecyclerViewAdapter
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -26,18 +22,6 @@ class AllPokemonTabFragment : Fragment() {
     private var _binding: FragmentTabAllPokemonBinding? = null
     private val binding get() = _binding!!
 
-    private lateinit var linearLayoutManager: LinearLayoutManager
-    private lateinit var gridLayoutManager: GridLayoutManager
-
-    private val pokemonListUpdatedObserver =
-        Observer<List<PokemonList>> { listOfPokemon ->
-            binding.allPokemonTabRecyclerview.adapter =
-                PokemonRecyclerViewAdapter(listOfPokemon) { name ->
-                    pokemonListItemClicked(
-                        name
-                    )
-                }
-        }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -47,7 +31,7 @@ class AllPokemonTabFragment : Fragment() {
         _binding = FragmentTabAllPokemonBinding.inflate(inflater, container, false)
 
         // Adding observers
-        allPokemonViewModel.listOfPokemon.observe(viewLifecycleOwner, pokemonListUpdatedObserver)
+      //  allPokemonViewModel.listOfPokemon.observe(viewLifecycleOwner, pokemonListUpdatedObserver)
 
         return binding.root
     }
@@ -55,14 +39,7 @@ class AllPokemonTabFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        allPokemonViewModel.getPokemon()
 
-        linearLayoutManager = LinearLayoutManager(context)
-        gridLayoutManager = GridLayoutManager(context, 2)
-
-        binding.apply {
-            allPokemonTabRecyclerview.layoutManager = gridLayoutManager
-        }
     }
 
     override fun onDestroyView() {
@@ -70,8 +47,4 @@ class AllPokemonTabFragment : Fragment() {
         _binding = null
     }
 
-    private fun pokemonListItemClicked(string: String) {
-        Log.d("test", string)
-       // NavHostFragment.findNavController(requireParentFragment()).navigate()
-    }
 }
