+++ app/src/main/java/com/minjee/favoritepokemon/ui/pokemon/PokemonViewModel.kt
@@ -1,17 +1,16 @@
-package com.minjee.favoritepokemon.ui.pokemon.allpokemon
+package com.minjee.favoritepokemon.ui.pokemon
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.minjee.favoritepokemon.data.PokemonList
 import com.minjee.favoritepokemon.data.PokemonRepository
 import com.minjee.favoritepokemon.data.PokemonRequest
-import com.minjee.favoritepokemon.data.Response
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class AllPokemonTabViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
+class PokemonViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
 
     val listOfPokemon: MutableLiveData<List<PokemonList>> = MutableLiveData()
