+++ app/src/main/res/layout/fragment_tab_my_pokemon.xml
@@ -4,7 +4,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".ui.pokemon.mypokemon.MyPokemonTabFragment">
+tools:context=".ui.favorites.mypokemon.MyPokemonTabFragment">
 
     <TextView
         android:layout_width="wrap_content"
